+++ deepinv/physics/blur.py
@@ -740,30 +740,29 @@ class TiledSpaceVaryingBlur(LinearPhysics):
         This class only supports `'valid'` padding. If you need other padding options, please raise an issue.
 
     |sep|
-    
+
     :Examples:
-    
-    >>> import torch 
+
+    >>> import torch
     >>> from deepinv.physics import TiledSpaceVaryingBlur
     >>> from deepinv.physics.generator import MotionBlurGenerator, TiledBlurGenerator
-    >>> import deepinv as dinv 
+    >>> import deepinv as dinv
     >>> img_size = (256, 256)
     >>> patch_size = (64, 64)
     >>> stride = (32, 32)
     >>> x = dinv.utils.load_example(
     ...        "butterfly.png", img_size=img_size, resize_mode="resize"
-    ...    ).to(device)
-    >>> psf_generator = DiffractionBlurGenerator(psf_size=(31, 31), device=device)
+    ...    )
+    >>> psf_generator = DiffractionBlurGenerator(psf_size=(31, 31))
     >>> generator = TiledBlurGenerator(
     ...     psf_generator=psf_generator,
     ...     patch_size=patch_size,
     ...     stride=stride,
-    ...     device=device,
     ... )
     >>> filters = generator.step(batch_size=1, img_size=img_size)["filters"]
-    >>> physics = TiledSpaceVaryingBlur(patch_size=patch_size, stride=stride, device=device)
+    >>> physics = TiledSpaceVaryingBlur(patch_size=patch_size, stride=stride)
     >>> y = physics(x, filters=filters)
-    >>> dinv.utils.plot([x, y], titles=["Original", "Blurred"])
+    >>> dinv.utils.plot([x, y], titles=["Original", "Blurred"])   # DOCTEST: +SKIP
     """
 
     def __init__(
@@ -779,9 +778,10 @@ class TiledSpaceVaryingBlur(LinearPhysics):
 
         self.patch_size = _as_pair(patch_size)
         self.stride = _as_pair(stride) if stride is not None else self.patch_size // 2
-        assert self.stride[0] <= self.patch_size[0] and self.stride[1] <= self.patch_size[1], (
-            f"Stride {self.stride} must be smaller or equal than patch_size {self.patch_size}."
-        )
+        assert (
+            self.stride[0] <= self.patch_size[0]
+            and self.stride[1] <= self.patch_size[1]
+        ), f"Stride {self.stride} must be smaller or equal than patch_size {self.patch_size}."
         self._dynamic_img_size = None  # To track image size changes
 
         self.use_fft = use_fft
@@ -879,7 +879,7 @@ class TiledSpaceVaryingBlur(LinearPhysics):
 
         :param torch.Tensor y: blurry image :math:`y`. Tensor of size `(B, C, H', W')`.
         :param torch.Tensor filters: Filters :math:`h_k`. Tensor of size `(b, c, K, h, w)`.
-        
+
         :return: torch.Tensor: Adjoint result.
         """
         if filters is None: