+++ deepinv/physics/generator/blur.py
@@ -1113,12 +1113,11 @@ class TiledBlurGenerator(PSFGenerator):
     Generates parameters of the :class:`deepinv.physics.TiledSpaceVaryingBlur` operator.
     The image is divided into overlapping patches, each local patch is convolved with a different PSF.
 
-    Parameters generated: `'filters'`: tensor of shape `(B, C, K, psf_size, psf_size)`
+    This generates a dict with key `'filter'`, which is tensor of shape `(B, C, K, psf_size, psf_size)`
     where `K` is the number of patches in which the image is divided.
     It is computed based on the `patch_size`, `stride` and the given `img_size` during the `step()` function call.
 
-    :param deepinv.physics.generator.PSFGenerator psf_generator: A PSF generator, such as :class:`motion blur <deepinv.physics.generator.MotionBlurGenerator>`
-    or :class:`diffraction blur generator <deepinv.physics.generator.DiffractionBlurGenerator>`
+    :param deepinv.physics.generator.PSFGenerator psf_generator: A PSF generator, such as :class:`motion blur <deepinv.physics.generator.MotionBlurGenerator>` or :class:`diffraction blur generator <deepinv.physics.generator.DiffractionBlurGenerator>`.
 
     :param int | tuple[int, int] patch_size: size of the patches (height, width) in which the image is divided.
     :param int | tuple[int, int] stride: stride between adjacent patches (height, width). Defaults to `patch_size`.